+++ client/allocrunner/taskrunner/getter/getter_test.go
@@ -2,7 +2,9 @@ package getter
 
 import (
 	"fmt"
+	"io"
 	"io/ioutil"
+	"mime"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -18,14 +20,92 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// fakeReplacer is a noop version of taskenv.TaskEnv.ReplaceEnv
-type fakeReplacer struct{}
+// noopReplacer is a noop version of taskenv.TaskEnv.ReplaceEnv.
+type noopReplacer struct{}
 
-func (fakeReplacer) ReplaceEnv(s string) string {
+func (noopReplacer) ReplaceEnv(s string) string {
 	return s
 }
 
-var taskEnv = fakeReplacer{}
+var noopTaskEnv = noopReplacer{}
+
+// upperReplacer is a version of taskenv.TaskEnv.ReplaceEnv that upper-cases
+// the given input.
+type upperReplacer struct{}
+
+func (upperReplacer) ReplaceEnv(s string) string {
+	return strings.ToUpper(s)
+}
+
+func removeAllT(t *testing.T, path string) {
+	require.NoError(t, os.RemoveAll(path))
+}
+
+func TestGetArtifact_getHeaders(t *testing.T) {
+	t.Run("nil", func(t *testing.T) {
+		require.Nil(t, getHeaders(noopTaskEnv, nil))
+	})
+
+	t.Run("empty", func(t *testing.T) {
+		require.Nil(t, getHeaders(noopTaskEnv, make(map[string]string)))
+	})
+
+	t.Run("set", func(t *testing.T) {
+		upperTaskEnv := new(upperReplacer)
+		expected := make(http.Header)
+		expected.Set("foo", "BAR")
+		result := getHeaders(upperTaskEnv, map[string]string{
+			"foo": "bar",
+		})
+		require.Equal(t, expected, result)
+	})
+}
+
+func TestGetArtifact_Headers(t *testing.T) {
+	file := "output.txt"
+
+	// Create the test server with a handler that will validate headers are set.
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Validate the expected value for our header.
+		value := r.Header.Get("X-Some-Value")
+		require.Equal(t, "FOOBAR", value)
+
+		// Write the value to the file that is our artifact, for fun.
+		w.Header().Set("Content-Type", mime.TypeByExtension(filepath.Ext(file)))
+		w.WriteHeader(http.StatusOK)
+		_, err := io.Copy(w, strings.NewReader(value))
+		require.NoError(t, err)
+	}))
+	defer ts.Close()
+
+	// Create a temp directory to download into.
+	taskDir, err := ioutil.TempDir("", "nomad-test")
+	require.NoError(t, err)
+	defer removeAllT(t, taskDir)
+
+	// Create the artifact.
+	artifact := &structs.TaskArtifact{
+		GetterSource: fmt.Sprintf("%s/%s", ts.URL, file),
+		GetterHeaders: map[string]string{
+			"X-Some-Value": "foobar",
+		},
+		RelativeDest: file,
+		GetterMode:   "file",
+	}
+
+	// Download the artifact.
+	taskEnv := new(upperReplacer)
+	err = GetArtifact(taskEnv, artifact, taskDir)
+	require.NoError(t, err)
+
+	// Verify artifact exists.
+	b, err := ioutil.ReadFile(filepath.Join(taskDir, file))
+	require.NoError(t, err)
+
+	// Verify we wrote the interpolated header value into the file that is our
+	// artifact.
+	require.Equal(t, "FOOBAR", string(b))
+}
 
 func TestGetArtifact_FileAndChecksum(t *testing.T) {
 	// Create the test server hosting the file to download
@@ -37,7 +117,7 @@ func TestGetArtifact_FileAndChecksum(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to make temp directory: %v", err)
 	}
-	defer os.RemoveAll(taskDir)
+	defer removeAllT(t, taskDir)
 
 	// Create the artifact
 	file := "test.sh"
@@ -49,7 +129,7 @@ func TestGetArtifact_FileAndChecksum(t *testing.T) {
 	}
 
 	// Download the artifact
-	if err := GetArtifact(taskEnv, artifact, taskDir); err != nil {
+	if err := GetArtifact(noopTaskEnv, artifact, taskDir); err != nil {
 		t.Fatalf("GetArtifact failed: %v", err)
 	}
 
@@ -69,7 +149,7 @@ func TestGetArtifact_File_RelativeDest(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to make temp directory: %v", err)
 	}
-	defer os.RemoveAll(taskDir)
+	defer removeAllT(t, taskDir)
 
 	// Create the artifact
 	file := "test.sh"
@@ -83,7 +163,7 @@ func TestGetArtifact_File_RelativeDest(t *testing.T) {
 	}
 
 	// Download the artifact
-	if err := GetArtifact(taskEnv, artifact, taskDir); err != nil {
+	if err := GetArtifact(noopTaskEnv, artifact, taskDir); err != nil {
 		t.Fatalf("GetArtifact failed: %v", err)
 	}
 
@@ -103,7 +183,7 @@ func TestGetArtifact_File_EscapeDest(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to make temp directory: %v", err)
 	}
-	defer os.RemoveAll(taskDir)
+	defer removeAllT(t, taskDir)
 
 	// Create the artifact
 	file := "test.sh"
@@ -117,7 +197,7 @@ func TestGetArtifact_File_EscapeDest(t *testing.T) {
 	}
 
 	// attempt to download the artifact
-	err = GetArtifact(taskEnv, artifact, taskDir)
+	err = GetArtifact(noopTaskEnv, artifact, taskDir)
 	if err == nil || !strings.Contains(err.Error(), "escapes") {
 		t.Fatalf("expected GetArtifact to disallow sandbox escape: %v", err)
 	}
@@ -155,7 +235,7 @@ func TestGetArtifact_InvalidChecksum(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to make temp directory: %v", err)
 	}
-	defer os.RemoveAll(taskDir)
+	defer removeAllT(t, taskDir)
 
 	// Create the artifact with an incorrect checksum
 	file := "test.sh"
@@ -167,7 +247,7 @@ func TestGetArtifact_InvalidChecksum(t *testing.T) {
 	}
 
 	// Download the artifact and expect an error
-	if err := GetArtifact(taskEnv, artifact, taskDir); err == nil {
+	if err := GetArtifact(noopTaskEnv, artifact, taskDir); err == nil {
 		t.Fatalf("GetArtifact should have failed")
 	}
 }
@@ -216,7 +296,7 @@ func TestGetArtifact_Archive(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to make temp directory: %v", err)
 	}
-	defer os.RemoveAll(taskDir)
+	defer removeAllT(t, taskDir)
 
 	create := map[string]string{
 		"exist/my.config": "to be replaced",
@@ -232,7 +312,7 @@ func TestGetArtifact_Archive(t *testing.T) {
 		},
 	}
 
-	if err := GetArtifact(taskEnv, artifact, taskDir); err != nil {
+	if err := GetArtifact(noopTaskEnv, artifact, taskDir); err != nil {
 		t.Fatalf("GetArtifact failed: %v", err)
 	}
 
@@ -255,7 +335,7 @@ func TestGetArtifact_Setuid(t *testing.T) {
 	// files that exist in the artifact to ensure they are overridden
 	taskDir, err := ioutil.TempDir("", "nomad-test")
 	require.NoError(t, err)
-	defer os.RemoveAll(taskDir)
+	defer removeAllT(t, taskDir)
 
 	file := "setuid.tgz"
 	artifact := &structs.TaskArtifact{
@@ -265,7 +345,7 @@ func TestGetArtifact_Setuid(t *testing.T) {
 		},
 	}
 
-	require.NoError(t, GetArtifact(taskEnv, artifact, taskDir))
+	require.NoError(t, GetArtifact(noopTaskEnv, artifact, taskDir))
 
 	var expected map[string]int
 
@@ -390,7 +470,7 @@ func TestGetGetterUrl_Queries(t *testing.T) {
 
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
-			act, err := getGetterUrl(taskEnv, c.artifact)
+			act, err := getGetterUrl(noopTaskEnv, c.artifact)
 			if err != nil {
 				t.Fatalf("want %q; got err %v", c.output, err)
 			} else if act != c.output {